Move Python wrapping code for onedim to onedim/pymod/
