+++ setup.py
@@ -67,6 +67,7 @@ def main():
         packages=[
             'modmesh',
             'modmesh.onedim',
+            'modmesh.app',
         ],
         ext_modules=[CMakeExtension("modmesh._modmesh")],
         cmdclass={'build_ext': cmake_build_ext},
